[fresharch] Made script more configurable
Made fresharch a bit more configurable due to the fact that I don't
always want to install all of the packages that fresharch is setup to
install. I also don't like having to open up the file before I start my
install with nano and edit each line (painfully) on it's own.